use Stream for Decrypt
